+++ CHANGELOG.md
@@ -14,6 +14,7 @@ CHANGELOG for Sulu
     * ENHANCEMENT #2121 [All]                 Cache composer cache dir and prefer dist downloads on Travis
     * ENHANCEMENT #2114 [All]                 Update ffmpeg bundle and lib
     * ENHANCEMENT #2116 [All]                 Made restart of jackrabbit between tests configureable
+    * ENHANCEMENT #2107 [WebsiteBundle]       Fixed portal redirect to local
     * BUGFIX      #2090 [MediaBundle]         Fixed fallback of media file-version meta
     * BUGFIX      #2092 [ContactBundle]       Fixed new contact when creating a new contact in the account
     * BUGFIX      #2103 [MediaBundle]         Fixed upload new version for media without thumbnail